+++ java_tools/configuration_definition/src/test/java/com/rusefi/test/ConfigFieldParserTest.java
@@ -82,7 +82,7 @@ public class ConfigFieldParserTest {
                 "custom ego_sensor_e 1 bits, S08, @OFFSET@, [0:1], @@ego_sensor_e_enum@@\n" +
                 "ego_sensor_e afr_type1;\n" +
                 "ego_sensor_e afr_type2;\n" +
-                "int8_t int\n" +
+                "int16_t int\n" +
                 "end_struct\n";
         ReaderState state = new ReaderState();
 
@@ -90,7 +90,7 @@ public class ConfigFieldParserTest {
         state.readBufferedReader(test, tsProjectConsumer);
         assertEquals("afr_type1 = bits, S08, 0, [0:1], \"BPSX\", \"Innovate\", \"14Point7\"\n" +
                 "afr_type2 = bits, S08, 1, [0:1], \"BPSX\", \"Innovate\", \"14Point7\"\n" +
-                "int = scalar, S08, 2, \"\", 1, 0, 0, 100, 0\n" +
+                "int = scalar, S16, 2, \"\", 1, 0, 0, 32000, 0\n" +
                 "; total TS size = 4\n", tsProjectConsumer.getContent());
     }