+++ firmware/CHANGELOG.md
@@ -48,6 +48,7 @@ Release template (copy/paste this for new release):
  - Fuel Priming reset fix #4627
  - Slower than expected RPM information was slowing engine start #4629
  - Fix 36-2-1 trigger (Mitsubishi 4B11, etc) #4635
+ - Improve injection phase table defaults #4808
 
 ### Removed
  - Idle timing deadzone #4729